+++ seedfarmer/commands/_bootstrap_commands.py
@@ -27,7 +27,6 @@ from jinja2 import Template
 from seedfarmer import CLI_ROOT
 from seedfarmer.services import create_new_session, get_account_id, get_region
 from seedfarmer.services._iam import get_role
-from seedfarmer.utils import CfnSafeYamlLoader
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -38,7 +37,7 @@ def get_toolchain_template(
     permissions_boundary_arn: Optional[str] = None,
 ) -> Dict[Any, Any]:
     with open((os.path.join(CLI_ROOT, "resources/toolchain_role.template")), "r") as f:
-        role = yaml.load(f, CfnSafeYamlLoader)
+        role = yaml.safe_load(f)
     if principal_arn:
         role["Resources"]["ToolchainRole"]["Properties"]["AssumeRolePolicyDocument"]["Statement"][0]["Principal"][
             "AWS"
@@ -54,7 +53,7 @@ def get_deployment_template(
     toolchain_account_id: str, project_name: str, permissions_boundary_arn: Optional[str] = None
 ) -> Dict[Any, Any]:
     with open((os.path.join(CLI_ROOT, "resources/deployment_role.template")), "r") as f:
-        role = yaml.load(f, CfnSafeYamlLoader)
+        role = yaml.safe_load(f)
     if permissions_boundary_arn:
         role["Resources"]["DeploymentRole"]["Properties"]["PermissionsBoundary"] = permissions_boundary_arn
     template = Template(json.dumps(role))